Shut yum up
Our output matters. Programs will read it and parse it. This is in fact
the sole purpose of nbyum: print a parser-friendly and stable output, so
that they don't have to screen-scrape the yum console output.

As such, we don't want yum to print random stuff. So shut up yum!

git-svn-id: https://10.8.18.16/nbrs/5.0/apps/nbyum/HEAD@20302 ef222f69-f2fc-0310-9845-a72751d35017