Catch all errors in the main command
Error handling is simple: we print the error message in a JSON
dictionary.

As such, it makes sense to just not handle exceptions and error cases in
the API, but just thrown them back up to the main script, which will
cath them all and print as appropriate.

Writing a new command becomes easier.

git-svn-id: https://10.8.18.16/nbrs/5.0/apps/nbyum/HEAD@20306 ef222f69-f2fc-0310-9845-a72751d35017