New decorator to ensure running as root
Several commands will need to make sure they have the required
privileges: they must be run as root.

Right now, we check deep inside the ``update`` code for the current uid,
so that ``check-update`` can be run by any user.

And we'd have to do that for every little piece of code requiring root
privileges to run, which means we'll forget some.

Now, we'll be able to simply decorate commands that need to be run as
root, and leave others alone, which should be much less error-prone.

git-svn-id: https://10.8.18.16/nbrs/5.0/apps/nbyum/HEAD@20374 ef222f69-f2fc-0310-9845-a72751d35017