info: Check apostrophes in package details
This is actually what prompted me to use json.dumps() everywhere, as I
was having trouble properly escaping the apostrophes in our output.
json.dumps() does that for us, in a way that is compliant with the JSON
specification, so that the caller will always be able to parse it.

git-svn-id: https://10.8.18.16/nbrs/5.0/apps/nbyum/HEAD@20567 ef222f69-f2fc-0310-9845-a72751d35017