install: Don't miss the actually requested package
So it appears that the Yum API uses the "i" (install) ts_state in a
transaction only for packages that are **installed** as part of an
**update** (the ``installonly_pkgs``).

But when installing a package, with an actual **install** operation,
those get the ts_state "u" (update) and are either installed as
dependencies or just the actually requested install.

As a result, before this commit, we were missing the requested packages
from the transaction summary (but we were already handling installdeps
properly).

Thank you Yum developers for your straightforward API.

git-svn-id: https://10.8.18.16/nbrs/5.0/apps/nbyum/HEAD@20643 ef222f69-f2fc-0310-9845-a72751d35017