install: Improve recap ordering
So, since packages requested to be installed with an **install**
operation have the ts_state "u" in the transaction, we need to be even
more clevererer than before to order the packages in the transaction
summary.

Oh well...

Note that this is the same code-path as for summaries of updates
transactions (and remove in the future), however it should keep the
order consistent with what we have now.

git-svn-id: https://10.8.18.16/nbrs/5.0/apps/nbyum/HEAD@20644 ef222f69-f2fc-0310-9845-a72751d35017