tests: Factorize the nbyum invocation and checking of stdout
I grew tired of always copy-pasting the same result = [json.loads(....
lines in each and every unit tests, so let's put that in the parent
TestCase class.

This is a huge win as it greatly simplifies writing unit tests. And
since they are already quite annoying to write (with the need to create
a repo for each test), every incentive to write more counts. :)

git-svn-id: https://10.8.18.16/nbrs/5.0/apps/nbyum/HEAD@20652 ef222f69-f2fc-0310-9845-a72751d35017