Allow interacting with the plugins
Yum has an incredibly complex plugin system, and a few great plugins.

We want to use some of them, but it seems that to do that we must
either:
1. make sure they absolutely always run (e.g setting `remove_always' to
true for the remove-with-leaves plugin)
2. use Yum's optparse and pass the required command line arguments

The former is unacceptable, as we might want to use Yum directly without
those plugins (e.g to fix a broken situation on a host).

The latter is unacceptable because of the hairiness of Yum's optparse,
and because we want nbyum to always use those plugins, so that having to
use mandatory command line arguments is a big annoyance.

So, instead, we simulate the command line passed to the plugins by
creating mock optparse instances which conveniently contain the
"options" we want to pass to the plugins.

And then Yum will happily run those plugins.

Of course, that's a horrible solution, but it's the best I could find.

git-svn-id: https://10.8.18.16/nbrs/5.0/apps/nbyum/HEAD@20696 ef222f69-f2fc-0310-9845-a72751d35017