Print the first argument of the lock owner
It can be interesting to know that yum is locked by an update, or the
installation of a security module, or...

This will usually be the first argument after the executable name (the
yum / nbyum subcommand mostly).

So let's try to get it, and if we found it, add it to the error message.

git-svn-id: https://10.8.18.16/nbrs/5.0/apps/nbyum/HEAD@21067 ef222f69-f2fc-0310-9845-a72751d35017