list: Special case the available security modules
In the case of security modules, we want to show the users which ones
are installed and which ones are "available", but the latter term has a
different meaning as for packages.

Available security modules should be the ones that are not installed,
even if the installed security module is in a different version from the
one available (case in point: updates).

The idea is to provide the user with a simple information: did I
enable (and thus obtain a license) for a security module or not?

git-svn-id: https://10.8.18.16/nbrs/5.0/apps/nbyum/HEAD@21236 ef222f69-f2fc-0310-9845-a72751d35017