Fix the auto-update-debug-info Yum plugin
Some parts of Yum, as well as some plugins, seem to assume
things they probably shouldn't about the Yum command line.

Since we are faking the command line options with our own
DummyOpts class, we can break those assumptions.

The first issue to arise was with the auto-update-debug-info
plugin, which this commit fixes.

git-svn-id: https://10.8.18.16/nbrs/5.0/apps/nbyum/HEAD@21297 ef222f69-f2fc-0310-9845-a72751d35017