New utils: get_version
We agreed that Epoch, Version and Release don't make sense for the
user. They are only details relevant to the packager or the distro
maker, but the user only cares about a global "version", made out
of the three combined.

This function helps us get this "version" of a package, so we display it
in a way that makes sense to the user.

git-svn-id: https://10.8.18.16/nbrs/5.0/apps/nbyum/HEAD@21523 ef222f69-f2fc-0310-9845-a72751d35017