Work around a yum bug
Basically, yum has the following known bug:
    https://bugzilla.redhat.com/show_bug.cgi?id=684686#c6

It's harmless, and should only ever occur when removing a package
while installroot is not /

We probably won't do that ever, except in the unit tests, where we
do that just about 100% of the time.

So this bug breaks our unit tests, as it adds one error message in
the nbyum output.

Since it's absolutely harmless, I decided to hide this specific
error message if we're running with installroot different from /

git-svn-id: https://10.8.18.16/nbrs/5.0/apps/nbyum/HEAD@21535 ef222f69-f2fc-0310-9845-a72751d35017