Log transactions to syslog, like Yum
Yum sends some messges (the ones from its RPM callback's filelog()
method) to syslog.

We want that, but we broke it with our custom logger.

So first, let's stop overwriting filelog(), so that the one in Yum
still sends its messages.

Then, when we intercept those messages in our custom logger, make
sure we send them to syslog.

git-svn-id: https://10.8.18.16/nbrs/5.0/apps/nbyum/HEAD@21687 ef222f69-f2fc-0310-9845-a72751d35017