Progress and recaps are not displayed on 32 bits
We want to use **big** numbers for the progress and recap
level, so that they will always be displayed, even if we set
the Yum logging stack to only print warn and above. Our
levels are above.

However, it seems I chose 2 too big numbers. On 64 bits, where
I did all my testing, those numbers are store by Python in an
integer, and on 32 bits they are stored in a long.

But that's just internal implementation details, Python developers
shouldn't have to care about that... Right?

Well, the logging module uses numbers to represent the levels,
so at some point, in its log() method, it checks that the level
is a number. So far, so good.

Except it doesn't check if the level is a number. It checks for
integers, and only integers. Not longs.

There, our levels were not integers on 32 bits, they were longs
so the logging module would refuse them and not print the
messages.

I have no idea why they made it this way, but I guess I just
have to use smaller numbers.

git-svn-id: https://10.8.18.16/nbrs/5.0/apps/nbyum/HEAD@21720 ef222f69-f2fc-0310-9845-a72751d35017