Don't raise an exception in the middle of the transaction
This would have stopped the transaction from going further, leaving
the system in a half-broken state.

We should simply log the error instead, as this is bad and we don't
know what to do with the package, but we certainly have to finish
the transaction at this point.

Also, the error message was rephrased, to make it more helpful.

git-svn-id: https://10.8.18.16/nbrs/5.0/apps/nbyum/HEAD@21730 ef222f69-f2fc-0310-9845-a72751d35017