update: Suggest a reboot on some package updates
Packages like the kernel (urrently the only one hardcoded) will always
require a reboot after an update before the new code is used.

We don't want to force the update at an inconvenient time, but we need
to remind the administrator that they should schedule a reboot at some
point.

git-svn-id: https://10.8.18.16/nbrs/5.0/apps/nbyum/HEAD@25590 ef222f69-f2fc-0310-9845-a72751d35017