yumbase: Access the config early
Under the hood, accessing the config calls the base._getConfig() method,
which itselfs does a bunch of preparation stuff, like reading the
configs, but also... setting up the logging.

That seems like a good idea to do early on, especially as we'll soon
(i.e next commit) need to do some logging very early on.