list: Rework the hiding feature
With the previous version, we found a problem with the following
scenario:

  - we had two versions of a given package
  - older version was not hidden, newer was

Because we were filtering on the hidden status before grouping by name
and getting the latest one:

  - with --show-hidden, nbyum would display the latest version
  - without --show-hidden, nbyum would display the oldest one

This is obviously not what we wanted. ;)

The solution is to filter the hidden one out **after** having computed
the latest version (what Yum calls the "best" package).