utils: New locking decorator
This is heavily inspired by our current locking context manager, with
one difference: this checks if the current process is the lock owner,
and if yes then it doesn't complain.

Using it as a decorator allows offloading it to the utils module,
instead of having it in the main class.

Also, it makes the main class much more readable.

Finally, it is much easier to lock a method which wasn't locked than
with the previous way.